removing the 'analysis' property from 'EvidenceReference'.
(This action was done unilaterally because we didn't have time to discuss it with the community before cutting milestone 1. The removal was done preemptively because removal at a later date would consitute a backwards-incompatible change. We will open up a separate issue to discuss the change and add it back later if it's determined that it's needed. As such, an addition will be a backwards-compatible change.)